requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.11.3
pydantic-settings==2.3.0
pydantic_core==2.33.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
cryptography==42.0.8
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
python-jose==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
passlib==1.7.4
bcrypt==4.3.0
dnspython==2.7.0
